First Functional Scrapes
Prints page source. Scanning this for appropriate scrapcomming information.

Commented out the wait.until function

As far as I understand the wait.until function because that part of scraping that I didn't need. Which seems superfluous now, but as soon as I commented it out I got output to my sqlite DB.